// rpg-tactics/src/client/sprite-metadata.ts
// Generated by npm run build:sprites.
export const SPRITE_META = {
  "hellhound": {
    "span": 242.64018518541025,
    "center": 51.90757350594606
  },
  "hellhound-dead": {
    "span": 242.64018518541025,
    "center": 27.269924299472358
  },
  "player": {
    "span": 118.73757916473309,
    "center": 25.801318615826432
  },
  "player-bite": {
    "span": 118.73757916473309,
    "center": 25.801318615826432
  },
  "player-eat": {
    "span": 118.73757916473309,
    "center": 25.801318615826432
  },
  "player-dead": {
    "span": 118.73757916473309,
    "center": 13.984576563831975
  },
  "bat": {
    "span": 374.40000000000003,
    "center": 0
  },
  "bat-dead": {
    "span": 374.40000000000003,
    "center": 12.703488160350366
  },
  "spider": {
    "span": 252.28125000000009,
    "center": 22.083171773509108
  },
  "spider-wall": {
    "span": 252.28125000000009,
    "center": 22.083171773509108
  },
  "spider-dead": {
    "span": 252.28125000000003,
    "center": 22.083171773509097
  },
  "gargoyle": {
    "span": 117.88497933523448,
    "center": 39.74999999999999
  },
  "gargoyle-dead": {
    "span": 117.88497933523448,
    "center": 38.52428438369427
  },
  "cross": {
    "span": 70.19999999999999,
    "center": 26.999999999999996
  },
  "boulder": {
    "span": 105.30000000000001,
    "center": 27.701743897835982
  },
  "angel": {
    "span": 163.79999999999998,
    "center": 62.999999999999986
  },
  "gem": {
    "span": 60.45000000000002,
    "center": 14.470243197438984
  }
} as const;